perf(support): remove the elvis-support crate
