+++ crates/core/Cargo.toml
@@ -13,7 +13,7 @@ keywords = ["js", "wasm", "spa", "web"]
 readme = './README.md'
 
 [dependencies]
-elvis-core-support = { path = "../core-support" }
+elvis-core-support = { path = "../core-support", version = "0.1.0" }
 
 [features]
 web = []
