+++ crates/derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "elvis-derive"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["clearloop <udtrokia@gmail.com>"]
 edition = "2018"
 license = "MIT/Apache-2.0"
