+++ crates/epm/src/template.rs
@@ -11,7 +11,7 @@ content = [77, 73, 84, 32, 76, 105, 99, 101, 110, 115, 101, 10, 10, 67, 111, 112
 
 [[children]]
 path = "/var/folders/5_/m9p8c1mn6k34_x77p79v15z00000gn/T/hello-world/README.md"
-content = [35, 32, 69, 108, 118, 105, 115, 32, 65, 80, 80, 10, 10, 84, 104, 105, 115, 32, 112, 114, 111, 106, 101, 99, 116, 32, 105, 115, 32, 98, 111, 111, 116, 115, 116, 114, 97, 112, 112, 101, 100, 32, 98, 121, 32, 91, 101, 112, 109, 93, 91, 49, 93, 46, 10, 10, 96, 96, 96, 10, 99, 97, 114, 103, 111, 32, 105, 110, 115, 116, 97, 108, 108, 32, 101, 112, 109, 10, 96, 96, 96, 10, 10, 70, 105, 110, 100, 32, 116, 104, 101, 32, 109, 111, 115, 116, 32, 114, 101, 99, 101, 110, 116, 32, 118, 101, 114, 115, 105, 111, 110, 32, 111, 102, 32, 116, 104, 105, 115, 32, 103, 117, 105, 100, 101, 32, 97, 116, 32, 91, 104, 101, 114, 101, 93, 91, 50, 93, 46, 32, 65, 110, 100, 32, 99, 104, 101, 99, 107, 32, 111, 117, 116, 32, 91, 101, 108, 118, 105, 115, 93, 91, 51, 93, 32, 102, 111, 114, 32, 116, 104, 101, 32, 109, 111, 115, 116, 32, 117, 112, 45, 116, 111, 45, 100, 97, 116, 101, 32, 105, 110, 102, 111, 46, 10, 10, 10, 35, 35, 32, 65, 118, 97, 105, 108, 97, 98, 108, 101, 32, 83, 99, 114, 105, 112, 116, 115, 10, 10, 35, 35, 35, 32, 101, 112, 109, 32, 100, 101, 118, 10, 10, 82, 117, 110, 115, 32, 116, 104, 101, 32, 97, 112, 112, 32, 105, 110, 32, 116, 104, 101, 32, 100, 101, 118, 101, 108, 111, 112, 109, 101, 110, 116, 32, 109, 111, 100, 101, 46, 10, 79, 112, 101, 110, 32, 104, 116, 116, 112, 58, 47, 47, 108, 111, 99, 97, 108, 104, 111, 115, 116, 58, 51, 48, 48, 48, 32, 116, 111, 32, 118, 105, 101, 119, 32, 105, 116, 32, 105, 110, 32, 116, 104, 101, 32, 98, 114, 111, 119, 115, 101, 114, 46, 10, 10, 84, 104, 101, 32, 112, 97, 103, 101, 32, 119, 105, 108, 108, 32, 114, 101, 108, 111, 97, 100, 32, 105, 102, 32, 121, 111, 117, 32, 109, 97, 107, 101, 32, 101, 100, 105, 116, 115, 46, 10, 89, 111, 117, 32, 119, 105, 108, 108, 32, 97, 108, 115, 111, 32, 115, 101, 101, 32, 97, 110, 121, 32, 101, 114, 114, 111, 114, 115, 32, 105, 110, 32, 116, 104, 101, 32, 99, 111, 110, 115, 111, 108, 101, 46, 10, 10, 35, 35, 35, 32, 101, 112, 109, 32, 98, 117, 105, 108, 100, 10, 10, 73, 116, 32, 99, 111, 114, 114, 101, 99, 116, 108, 121, 32, 98, 117, 110, 100, 108, 101, 115, 32, 69, 108, 118, 105, 115, 32, 105, 110, 32, 112, 114, 111, 100, 117, 99, 116, 105, 111, 110, 32, 109, 111, 100, 101, 32, 97, 110, 100, 32, 111, 112, 116, 105, 109, 105, 122, 101, 115, 32, 116, 104, 101, 32, 98, 117, 105, 108, 100, 32, 102, 111, 114, 32, 116, 104, 101, 32, 98, 101, 115, 116, 32, 112, 101, 114, 102, 111, 114, 109, 97, 110, 99, 101, 46, 10, 10, 10, 35, 35, 32, 83, 111, 109, 101, 116, 104, 105, 110, 103, 32, 77, 105, 115, 115, 105, 110, 103, 63, 10, 10, 68, 111, 110, 92, 96, 116, 32, 84, 104, 105, 110, 107, 32, 84, 119, 105, 99, 101, 44, 32, 105, 116, 92, 96, 115, 32, 65, 108, 114, 105, 103, 104, 116, 46, 10, 10, 91, 49, 93, 58, 32, 104, 116, 116, 112, 115, 58, 47, 47, 103, 105, 116, 104, 117, 98, 46, 99, 111, 109, 47, 101, 108, 118, 105, 115, 106, 115, 47, 101, 108, 118, 105, 115, 47, 116, 114, 101, 101, 47, 109, 97, 115, 116, 101, 114, 47, 99, 114, 97, 116, 101, 115, 47, 101, 112, 109, 10, 91, 50, 93, 58, 32, 104, 116, 116, 112, 115, 58, 47, 47, 101, 108, 118, 105, 115, 106, 115, 46, 103, 105, 116, 104, 117, 98, 46, 105, 111, 47, 98, 111, 111, 107, 10, 91, 51, 93, 58, 32, 104, 116, 116, 112, 115, 58, 47, 47, 103, 105, 116, 104, 117, 98, 46, 99, 111, 109, 47, 101, 108, 118, 105, 115, 106, 115, 47, 101, 108, 118, 105, 115, 10]
+content = [35, 32, 69, 108, 118, 105, 115, 32, 65, 80, 80, 10, 10, 84, 104, 105, 115, 32, 112, 114, 111, 106, 101, 99, 116, 32, 105, 115, 32, 98, 111, 111, 116, 115, 116, 114, 97, 112, 112, 101, 100, 32, 98, 121, 32, 91, 101, 112, 109, 93, 91, 49, 93, 46, 10, 10, 96, 96, 96, 10, 99, 97, 114, 103, 111, 32, 105, 110, 115, 116, 97, 108, 108, 32, 101, 112, 109, 10, 96, 96, 96, 10, 10, 70, 105, 110, 100, 32, 116, 104, 101, 32, 109, 111, 115, 116, 32, 114, 101, 99, 101, 110, 116, 32, 118, 101, 114, 115, 105, 111, 110, 32, 111, 102, 32, 116, 104, 105, 115, 32, 103, 117, 105, 100, 101, 32, 97, 116, 32, 91, 104, 101, 114, 101, 93, 91, 50, 93, 46, 32, 65, 110, 100, 32, 99, 104, 101, 99, 107, 32, 111, 117, 116, 32, 91, 101, 108, 118, 105, 115, 93, 91, 51, 93, 32, 102, 111, 114, 32, 116, 104, 101, 32, 109, 111, 115, 116, 32, 117, 112, 45, 116, 111, 45, 100, 97, 116, 101, 32, 105, 110, 102, 111, 46, 10, 10, 10, 35, 35, 32, 65, 118, 97, 105, 108, 97, 98, 108, 101, 32, 83, 99, 114, 105, 112, 116, 115, 10, 10, 35, 35, 35, 32, 96, 101, 112, 109, 32, 100, 101, 118, 96, 10, 10, 82, 117, 110, 115, 32, 116, 104, 101, 32, 97, 112, 112, 32, 105, 110, 32, 116, 104, 101, 32, 100, 101, 118, 101, 108, 111, 112, 109, 101, 110, 116, 32, 109, 111, 100, 101, 46, 10, 79, 112, 101, 110, 32, 104, 116, 116, 112, 58, 47, 47, 108, 111, 99, 97, 108, 104, 111, 115, 116, 58, 51, 48, 48, 48, 32, 116, 111, 32, 118, 105, 101, 119, 32, 105, 116, 32, 105, 110, 32, 116, 104, 101, 32, 98, 114, 111, 119, 115, 101, 114, 46, 10, 10, 84, 104, 101, 32, 112, 97, 103, 101, 32, 119, 105, 108, 108, 32, 114, 101, 108, 111, 97, 100, 32, 105, 102, 32, 121, 111, 117, 32, 109, 97, 107, 101, 32, 101, 100, 105, 116, 115, 46, 10, 89, 111, 117, 32, 119, 105, 108, 108, 32, 97, 108, 115, 111, 32, 115, 101, 101, 32, 97, 110, 121, 32, 101, 114, 114, 111, 114, 115, 32, 105, 110, 32, 116, 104, 101, 32, 99, 111, 110, 115, 111, 108, 101, 46, 10, 10, 35, 35, 35, 32, 96, 101, 112, 109, 32, 98, 117, 105, 108, 100, 96, 10, 10, 73, 116, 32, 99, 111, 114, 114, 101, 99, 116, 108, 121, 32, 98, 117, 110, 100, 108, 101, 115, 32, 69, 108, 118, 105, 115, 32, 105, 110, 32, 112, 114, 111, 100, 117, 99, 116, 105, 111, 110, 32, 109, 111, 100, 101, 32, 97, 110, 100, 32, 111, 112, 116, 105, 109, 105, 122, 101, 115, 32, 116, 104, 101, 32, 98, 117, 105, 108, 100, 32, 102, 111, 114, 32, 116, 104, 101, 32, 98, 101, 115, 116, 32, 112, 101, 114, 102, 111, 114, 109, 97, 110, 99, 101, 46, 10, 10, 10, 35, 35, 32, 83, 111, 109, 101, 116, 104, 105, 110, 103, 32, 77, 105, 115, 115, 105, 110, 103, 63, 10, 10, 68, 111, 110, 92, 96, 116, 32, 84, 104, 105, 110, 107, 32, 84, 119, 105, 99, 101, 44, 32, 105, 116, 92, 96, 115, 32, 65, 108, 114, 105, 103, 104, 116, 46, 10, 10, 91, 49, 93, 58, 32, 104, 116, 116, 112, 115, 58, 47, 47, 103, 105, 116, 104, 117, 98, 46, 99, 111, 109, 47, 101, 108, 118, 105, 115, 106, 115, 47, 101, 108, 118, 105, 115, 47, 116, 114, 101, 101, 47, 109, 97, 115, 116, 101, 114, 47, 99, 114, 97, 116, 101, 115, 47, 101, 112, 109, 10, 91, 50, 93, 58, 32, 104, 116, 116, 112, 115, 58, 47, 47, 101, 108, 118, 105, 115, 106, 115, 46, 103, 105, 116, 104, 117, 98, 46, 105, 111, 47, 98, 111, 111, 107, 10, 91, 51, 93, 58, 32, 104, 116, 116, 112, 115, 58, 47, 47, 103, 105, 116, 104, 117, 98, 46, 99, 111, 109, 47, 101, 108, 118, 105, 115, 106, 115, 47, 101, 108, 118, 105, 115, 10]
 
 [[children]]
 path = "/var/folders/5_/m9p8c1mn6k34_x77p79v15z00000gn/T/hello-world/lib.rs"
