--- crates/support/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "elvis-support"
-version = "0.1.0"
-authors = ["clearloop <udtrokia@gmail.com>"]
-edition = "2018"
-license = "MIT/Apache-2.0"
-repository = "https://github.com/elvisjs/elvis/tree/master/crates/support"
-homepage = "https://github.com/elvisjs"
-documentation = "https://docs.rs/elvis-support"
-description = """
-An internal dependency of elvis
-"""
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1.0.7"
-syn = "1.0.33"
